Renaming betterversion.better to be projects.version_handling
